style(dropdown): add options to components
