+++ packages/dropdown/src/Dropdown.types.ts
@@ -1,23 +1,6 @@
 import React from 'react';
 
-export type DropdownOptionProps = {
-  /**
-   * labels of options
-   */
-  label: string;
-  /**
-   * value of options
-   */
-  value: string;
-  /**
-   * Element placed before the children.
-   */
-  leftContent?: React.ReactNode;
-  /**
-   * Element placed after the children
-   */
-  rightContent?: React.ReactNode;
-};
+import { DropdownOptionProps } from './components/DropdownOption.types';
 
 export type DropdownProps = {
   /**
