+++ packages/dropdown/src/components/DropdownOption.types.ts
@@ -1,0 +1,20 @@
+import React from 'react';
+
+export type DropdownOptionProps = {
+  /**
+   * labels of options
+   */
+  label: string;
+  /**
+   * value of options
+   */
+  value: string;
+  /**
+   * Element placed before the children.
+   */
+  leftContent?: React.ReactNode;
+  /**
+   * Element placed after the children
+   */
+  rightContent?: React.ReactNode;
+};
